pinctrl: nexell: protect irq_chip operations with spinlocks
struct nexell_pin_bank already contains a spinlock which
seems to have been designed to protect operations on pin banks.
However, this is never used anywhere in the code, appearing as
if it has just been forgotten.
Just use it around dangerous operations which might be subject
to race conditions.
While at it, add a DMB (data memory barrier) after irq_gpio_mask(),
like its closest friend irq_gpio_unmask().

BACKGROUND:
An issue has been observed in a particular case with two interrupts
on the same GPIO bank, both configured with threaded handlers
(IRQF_ONESHOT). In this scenario, the IRQ subsystem will mask the
interrupt in irq context and schedule the threaded handler.
Upon termination of the handler in thread context, the IRQ will be
unmasked again.
What has been observed is that *sometimes*, if IRQ2 is raised
*just about* at the same time the threaded handler for IRQ1 has
terminated (so it's about to get unmaked again), the handler for
IRQ1 will never be run again, appearing as if it never got unmasked.
So it looks like a race condition between the code running in IRQ
context masking IRQ2 and the one in thread context unmasking IRQ1.

This patch fixes the issue.

Signed-off-by: Gerlando Falauto <gerlando.falauto@gmail.com>
Cc: Bon-gyu, KOO <freestyle@nexell.co.kr>