pinctrl: nexell: use local flags for spin_lock_irqsave()/IO_LOCK()
this driver uses a spinlock for each GPIO module and spin_lock_irqsave().
spin_lock_irqsave() is a macro meant to save interrupt enable flags in
a *local* unsigned long variable, so that the same code can be executed
both in hardirq and softirq contexts.
Having a *static* variable to hold this information looks like a bad idea.
So use an automatic flags variable like everyone else does.

BACKGROUND:
An issue has been observed in a particular case with two interrupts
on the same GPIO bank, both configured with threaded handlers
(IRQF_ONESHOT). In this scenario, the IRQ subsystem will mask the
interrupt in irq context and schedule the threaded handler.
Upon termination of the handler in thread context, the IRQ will be
unmasked again.
What has been observed is that *sometimes*, if IRQ2 is raised
*just about* at the same time the threaded handler for IRQ1 has
terminated (so it's about to get unmaked again), the handler for
IRQ1 will never be run again, appearing as if it never got unmasked.
So it looks like a race condition between the code running in IRQ
context masking IRQ2 and the one in thread context unmasking IRQ1.

NOTE: This patch *does* not fix this issue, though it may fix
other subtle issues.

Signed-off-by: Gerlando Falauto <gerlando.falauto@gmail.com>
Cc: Bon-gyu, KOO <freestyle@nexell.co.kr>